day_2/task_2/requirements.txt
# requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
python-docx
faiss-cpu
sentence-transformers
numpy